targets/nrf53xx: i2c: implment basic i2c-slave
This commit adds a basic I2C slave implementation based on the NRFX TWIS
driver. This implementation is heavily inspired by the one from nrf52xx.
There are some major differences though, as the I2C master role uses
TWIM for nrf53xx and TWI directly for nrf52xx.

Additionally this commit slightly changes the I2C master implementation
to initialize the TWIM driver when a [read,write] happens and the device
is not already running. Previously it would always initialize the TWIM
device immediately, leaving no time to configure the device as a slave
role. The deferred master init is also present in the current
implementation for nrf52xx and I haven't seen any issues with it.

Interfacing with the I2C slave has some rough edges. The user has to
reach into the internal structure of the i2c configuration to assign
callbacks. We should consider extending the MBED i2c_slave API in the
future. To assign callbacks and a client context, the
`i2c_slave_callbacks` struct should be configured within the i2c_s
structure.

Currently the implementation forces async mode using EasyDMA (forced by
the NRFX TWIS driver) which puts a requirement on buffers passed to the
i2c slave to be allocated in the `data` region of the memory.
The synchrounous calls to i2c_slave_[read,write] haven't been tested,
but they are copied from the nrf52xx implementation.

Another limitation of the implementation is that you may only use a
single I2C slave device due to the single global i2c_s pointer design.
Implementing more flexibility is definitely relevant, but is not that
straight forward to do since the TWIS driver API doesn't contain any
context passing.