nrf53xx: analogin: prevent errornous conversion
Debugging some unexplainably high voltage readouts in a real world
application, it was discovered that the nrf5340 ADC can (and will)
return negative values.
Occasionally, these would cause the conversion with lots of casting
to fail and the result would underflow.

This is a simple workaround for the problem. Since the function is
supposed to return a uint16_t it doesn't seem like an issue to clamp the
ADC reading positive to me.
It is clear that any applications relying on the `analogin_read_u16`
functionality must've worked with strictly positive ADC measurements as
otherwise, the applications would have been unreliable and return
incorrect measurements anyway.

Signed-off-by: Emil-Juhl <emdj@bang-olufsen.dk>